Rename plugins.v1 module to plugins.types
