+++ conda-store-server/conda_store_server/app.py
@@ -29,7 +29,7 @@ from traitlets.config import LoggingConfigurable
 from conda_store_server import CONDA_STORE_DIR, BuildKey, api, registry, storage
 from conda_store_server._internal import conda_utils, environment, orm, schema, utils
 from conda_store_server.plugins import hookspec, plugin_manager
-from conda_store_server.plugins.v1 import lock
+from conda_store_server.plugins.types import lock
 
 
 def conda_store_validate_specification(
